_articles/2025-09-28-reproducible-flax-orbax-grain/pyproject.toml
[project]
name = "2025-09-28-reproducible-flax-orbax-grain"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "flax>=0.12.0",
    "grain>=0.2.12",
    "optax>=0.2.6",
    "orbax-checkpoint>=0.11.25",
    "ruff>=0.13.2",
    "tensorflow>=2.20.0",
    "tensorflow-datasets>=4.9.9",
    "tqdm>=4.67.1",
]
